codestyle corrected

+++ Task_1_2_2/src/main/java/ru/nsu/pereverzev/HashTable.java
@@ -1,23 +1,38 @@
 package ru.nsu.pereverzev;
 
-import java.util.*;
-
 import static java.lang.Math.abs;
 
-public class HashTable <K,V> implements Iterable<Pair<K, V>>{
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.LinkedList;
+import java.util.ConcurrentModificationException;
+import java.util.List;
+import java.util.Iterator;
+
+/**
+ * hash table class.
+ */
+public class HashTable<K, V> implements Iterable<Pair<K, V>> {
     int elcnt;
     Semaphore semaphore;
     ArrayList<LinkedList<Pair<K, V>>> table;
+
     HashTable() {
         elcnt = 0;
         semaphore = new Semaphore();
         table = new ArrayList<>(Collections.nCopies(4, null));
     }
+
+    /**
+     * adding new element to the hash table.
+     */
     public void add(K key, V value) {
+        if(semaphore.isIterating()) {
+            throw new ConcurrentModificationException("trying to add while iterating");
+        }
         elcnt++;
-        if(elcnt > table.size() / 2) {
-            ArrayList<LinkedList<Pair<K, V>>> newTable = new ArrayList<>(
-                    Collections.nCopies(table.size() * 2, null));
+        if (elcnt > table.size() / 2) {
+            ArrayList<LinkedList<Pair<K, V>>> newTable = new ArrayList<>(Collections.nCopies(table.size() * 2, null));
             for (Pair<K, V> pair : this) {
                 int id = pair.getKey().hashCode() % newTable.size();
                 LinkedList<Pair<K, V>> bucket = newTable.get(id);
@@ -42,14 +57,20 @@ public class HashTable <K,V> implements Iterable<Pair<K, V>>{
         }
     }
 
+    /**
+     * remove element by key and value.
+     */
     public void remove(K key, V value) {
+        if(semaphore.isIterating()) {
+            throw new ConcurrentModificationException("trying to remove while iterating");
+        }
         int id = abs(abs(key.hashCode())) % table.size();
         LinkedList<Pair<K, V>> bucket = table.get(id);
         if (bucket == null) {
             throw new HashTableException("no such element to remove from table");
         } else {
             Pair<K, V> pair = new Pair<K, V>(key, value);
-            if(bucket.contains(pair)){
+            if (bucket.contains(pair)) {
                 bucket.removeLastOccurrence(pair);
             } else {
                 throw new HashTableException("no such element to remove from table");
@@ -57,6 +78,9 @@ public class HashTable <K,V> implements Iterable<Pair<K, V>>{
         }
     }
 
+    /**
+     * get value by the key.
+     */
     public V getValue(K key) {
         int id = abs(key.hashCode()) % table.size();
         if (table.get(id) == null) {
@@ -71,7 +95,13 @@ public class HashTable <K,V> implements Iterable<Pair<K, V>>{
         }
     }
 
+    /**
+     * update value founded by the key.
+     */
     public void updateValue(K key, V value) {
+        if(semaphore.isIterating()) {
+            throw new ConcurrentModificationException("trying to update while iterating");
+        }
         int id = abs(key.hashCode()) % table.size();
         List<Pair<K, V>> bucket = table.get(id);
         if (bucket == null) {
@@ -89,6 +119,9 @@ public class HashTable <K,V> implements Iterable<Pair<K, V>>{
         }
     }
 
+    /**
+     * returns true if there exist value with this key in table.
+     */
     public boolean existValue(K key) {
         int id = abs(key.hashCode()) % table.size();
         List<Pair<K, V>> bucket = table.get(id);
@@ -109,11 +142,14 @@ public class HashTable <K,V> implements Iterable<Pair<K, V>>{
         return new HashTableIterator<>(table, semaphore);
     }
 
+    /**
+     * method that converts table to string view.
+     */
     public String toString() {
         String str = "{";
         int cnt = 0;
         for (Pair<K, V> pair : this) {
-            if(cnt == 1) {
+            if (cnt == 1) {
                 str = str.concat(", ");
             }
             str = str.concat("'" + pair.getKey().toString() + "': " + pair.getVal().toString());
@@ -132,36 +168,47 @@ public class HashTable <K,V> implements Iterable<Pair<K, V>>{
             return false;
         }
         Iterator<Pair<K, V>> iter = iterator();
-        HashTable<K, V> tableThat = (HashTable<K, V>)obj;
+        HashTable<K, V> tableThat = (HashTable<K, V>) obj;
         while (iter.hasNext()) {
             Pair<K, V> pair = iter.next();
-            if(!tableThat.existValue(pair.getKey())) {
+            if (!tableThat.existValue(pair.getKey())) {
                 return false;
             }
-            if(tableThat.getValue(pair.getKey()) != pair.value) {
+            if (tableThat.getValue(pair.getKey()) != pair.value) {
                 return false;
             }
         }
         iter = tableThat.iterator();
         while (iter.hasNext()) {
             Pair<K, V> pair = iter.next();
-            if(!this.existValue(pair.getKey())) {
+            if (!this.existValue(pair.getKey())) {
                 return false;
             }
-            if(this.getValue(pair.getKey()) != pair.value) {
+            if (this.getValue(pair.getKey()) != pair.value) {
                 return false;
             }
         }
         return true;
     }
+
+    /**
+     * class that is used for detecting concurrent modification.
+     */
     public class Semaphore {
         int iterProcessCount;
+
         Semaphore() {
             iterProcessCount = 0;
         }
+
+        public boolean isIterating() {
+            return iterProcessCount > 0;
+        }
+
         public void increment() {
             iterProcessCount++;
         }
+
         public void decrement() {
             iterProcessCount--;
         }
